auth/requirements.txt
psycopg2
flask-sqlalchemy
flask-migrate
flask-failsafe